+++ src/data/eval_dataset/vidore_dataset.py
@@ -6,11 +6,11 @@ from src.utils.dataset_utils import load_hf_dataset, sample_dataset, load_qrels_
 from src.model.processor import process_input_text
 from src.utils.basic_utils import print_master
 
-# TASK_INST_QRY = "Find a document image that matches the given query:"
-# TASK_INST_TGT = "Understand the content of the provided document image."
+TASK_INST_QRY = "Find a document image that matches the given query:"
+TASK_INST_TGT = "Understand the content of the provided document image."
 # ColPali models use no prompts
-TASK_INST_QRY = ""
-TASK_INST_TGT = ""
+# TASK_INST_QRY = ""
+# TASK_INST_TGT = ""
 
 @add_metainfo_hook
 def data_prepare(batch_dict, **kwargs):