+++ CHANGELOG.md
@@ -6,6 +6,9 @@ All notable changes to this project will be documented in this file.
 [new-feature]: https://img.shields.io/badge/NEW%20FEATURE-brightgreen
 [new-release]: https://img.shields.io/badge/NEW%20RELEASE-orange
 
+## ![bug-fix] 2025-11-03
+- Fixed the issue in ```ViDoSeek-page``` and ```MMLongBench-page```. More details in this [issue](https://github.com/TIGER-AI-Lab/VLM2Vec/issues/167).
+
 ## ![new-release] 2025-08-08
 - Release the raw video files [here](https://huggingface.co/datasets/TIGER-Lab/MMEB_Raw_Video). Please note that raw videos are not required for MMEB evaluation — video frames are all you need. We provide the raw videos only in case they are useful for your own purposes. Please refer to our [main data repository](https://huggingface.co/datasets/TIGER-Lab/MMEB-V2) for more instructions.